+++ CHANGELOG.md
@@ -5,13 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.23.3] - 2020-12-14
+## [v0.23.3] - 2020-12-15
 
 ### Fixed
 
 - re-added missing fields from GetUsers query
   - family_name
   - given_name
+
+### Added
+
+- `MapIdToUser` components
+- UserPage at `/u` and `/u/*` 🚧 UNDER CONSTRUCTION
 
 ### Removed